Use latest symbol for global declarations
